backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI for LLM services
openai==1.12.0

# Fal.ai for image and video generation
fal-client==0.4.1

# ElevenLabs for music generation
elevenlabs==0.2.27

# PDF and text processing
PyPDF2==3.0.1
beautifulsoup4==4.12.3
requests==2.31.0

# Video editing (will be implemented)
moviepy==1.0.3

# Environment and utilities
python-dotenv==1.0.1
pydantic==2.5.3

# Async support
asyncio==3.4.3
aiofiles==23.2.1

